feat: 新增rabbitmq实时更新小说es索引信息功能

+++ src/main/java/io/github/novel/mynovel/dao/entity/BookInfo.java
@@ -32,7 +32,7 @@ public class BookInfo implements Serializable {
     /**
      * 作品方向;0-男频 1-女频
      */
-    private Byte workDirection;
+    private Integer workDirection;
 
     /**
      * 类别ID
